// src/modules/auth/auth.types.ts
export type RegisterAdminType = {
  workerId: string; // Ажилтны ID
  password: string; // Нууц үг
  surname: string; // Овог
  branch: string; // Салбарын ID
  givenname: string; // Нэр
  position: string; // Албан тушаал
  rank: string;
};

// RegisterAdminType jsondata